refactor: Default -> Simple 네이밍 변경
