+++ src/main/java/repository/SimpleCrudRepository.java
@@ -3,14 +3,14 @@ package repository;
 import java.util.List;
 import jdbc.JdbcTemplate;
 import persistence.dialect.Dialect;
-import persistence.entity.DefaultEntityManager;
+import persistence.entity.SimpleEntityManager;
 import persistence.entity.EntityManager;
 
-public class BaseCrudRepository<T, ID> extends AbstractRepository<T, ID> implements CrudRepository<T, ID> {
+public class SimpleCrudRepository<T, ID> extends AbstractRepository<T, ID> implements CrudRepository<T, ID> {
     private final EntityManager entityManager;
-    protected BaseCrudRepository(JdbcTemplate jdbcTemplate, Class<T> tClass, Dialect dialect) {
+    protected SimpleCrudRepository(JdbcTemplate jdbcTemplate, Class<T> tClass, Dialect dialect) {
         super(jdbcTemplate, tClass, dialect);
-        this.entityManager = new DefaultEntityManager(jdbcTemplate, entityMeta, dialect);
+        this.entityManager = new SimpleEntityManager(jdbcTemplate, entityMeta, dialect);
     }
 
     public T save(T entity) {
